api/requirements.txt
Flask
Flask-Mail
Flask-CORS
psycopg
python-dotenv
bcrypt
PyJWT